server/requirements.txt
flask
flask-cors
bcrypt
PyJWT
PyPDF2
psycopg2-binary
gunicorn
